+++ cryptofeed/backends/postgres.py
@@ -5,7 +5,7 @@ Please see the LICENSE file for the terms and conditions
 associated with this software.
 '''
 from datetime import datetime as dt
-import json
+from yapic import json
 
 import asyncpg