+++ cryptofeed/exchange/dsx.py
@@ -11,18 +11,18 @@ from decimal import Decimal
 from sortedcontainers import SortedDict as sd
 
 from cryptofeed.feed import Feed
-from cryptofeed.defines import TICKER, L2_BOOK, TRADES, BUY, SELL, BID, ASK, HITBTC
+from cryptofeed.defines import TICKER, L2_BOOK, TRADES, BUY, SELL, BID, ASK, DSX as DSX_id
 from cryptofeed.standards import pair_exchange_to_std, timestamp_normalize
 
 
 LOG = logging.getLogger('feedhandler')
 
 
-class HitBTC(Feed):
-    id = HITBTC
+class DSX(Feed):
+    id = DSX_id
 
     def __init__(self, pairs=None, channels=None, callbacks=None, **kwargs):
-        super().__init__('wss://api.hitbtc.com/api/2/ws',
+        super().__init__('wss://api.dsxglobal.com/api/2/ws',
                          pairs=pairs,
                          channels=channels,
                          callbacks=callbacks,