+++ cryptofeed/exchange/poloniex.py
@@ -4,7 +4,7 @@ Copyright (C) 2017-2020  Bryant Moscon - bmoscon@gmail.com
 Please see the LICENSE file for the terms and conditions
 associated with this software.
 '''
-import json
+from yapic import json
 import logging
 from decimal import Decimal
 import time
@@ -75,6 +75,9 @@ class Poloniex(Feed):
         # currencyPair, last, lowestAsk, highestBid, percentChange, baseVolume,
         # quoteVolume, isFrozen, 24hrHigh, 24hrLow
         pair_id, _, ask, bid, _, _, _, _, _, _ = msg
+        if pair_id not in self.pair_mapping:
+            # Ignore new trading pairs that are added during long running sessions
+            return
         pair = pair_exchange_to_std(self.pair_mapping[pair_id])
         if self.__do_callback(TICKER, pair):
             await self.callback(TICKER, feed=self.id,