+++ cryptofeed/exchanges.py
@@ -20,6 +20,7 @@ from cryptofeed.exchange.kraken_futures import KrakenFutures
 from cryptofeed.exchange.exx import EXX
 from cryptofeed.exchange.huobi import Huobi
 from cryptofeed.exchange.huobi_dm import HuobiDM
+from cryptofeed.exchange.huobi_swap import HuobiSwap
 from cryptofeed.exchange.okcoin import OKCoin
 from cryptofeed.exchange.okex import OKEx
 from cryptofeed.exchange.coinbene import Coinbene
@@ -31,3 +32,4 @@ from cryptofeed.exchange.bitcoincom import BitcoinCom
 from cryptofeed.exchange.bitmax import Bitmax
 from cryptofeed.exchange.upbit import Upbit
 from cryptofeed.exchange.blockchain import Blockchain
+from cryptofeed.exchange.dsx import DSX