+++ cryptofeed/feed.py
@@ -9,7 +9,7 @@ from collections import defaultdict
 
 from cryptofeed.callback import Callback
 from cryptofeed.standards import pair_std_to_exchange, feed_to_exchange, load_exchange_pair_mapping
-from cryptofeed.defines import TRADES, TICKER, L2_BOOK, L3_BOOK, VOLUME, FUNDING, BOOK_DELTA, OPEN_INTEREST, BID, ASK
+from cryptofeed.defines import TRADES, TICKER, L2_BOOK, L3_BOOK, VOLUME, FUNDING, BOOK_DELTA, OPEN_INTEREST, BID, ASK, LIQUIDATIONS
 from cryptofeed.util.book import book_delta, depth
 
 
@@ -54,8 +54,8 @@ class Feed:
                           VOLUME: Callback(None),
                           FUNDING: Callback(None),
                           OPEN_INTEREST: Callback(None),
-                          BOOK_DELTA: Callback(None)
-                          }
+                          BOOK_DELTA: Callback(None),
+                          LIQUIDATIONS: Callback(None)}
 
         if callbacks:
             for cb_type, cb_func in callbacks.items():