+++ cryptofeed/feedhandler.py
@@ -20,6 +20,7 @@ from cryptofeed.log import get_logger
 from cryptofeed.defines import DERIBIT, BINANCE, GEMINI, HITBTC, BITFINEX, BITMEX, BITSTAMP, POLONIEX, COINBASE, KRAKEN, KRAKEN_FUTURES, HUOBI, HUOBI_DM, OKCOIN, OKEX, COINBENE, BYBIT, BITTREX, BITCOINCOM, BINANCE_US, BITMAX, BINANCE_JERSEY, BINANCE_FUTURES, UPBIT
 from cryptofeed.defines import EXX as EXX_str
 from cryptofeed.defines import FTX as FTX_str
+from cryptofeed.defines import DSX as DSX_str
 from cryptofeed.exchanges import *
 from cryptofeed.nbbo import NBBO
 from cryptofeed.feed import RestFeed
@@ -57,7 +58,8 @@ _EXCHANGES = {
     BITTREX: Bittrex,
     BITCOINCOM: BitcoinCom,
     BITMAX: Bitmax,
-    UPBIT: Upbit
+    UPBIT: Upbit,
+    DSX_str: DSX
 }
 
 
@@ -91,7 +93,8 @@ class FeedHandler:
         timeout: int
             number of seconds without a message before the feed is considered
             to be timed out. The connection will be closed, and if retries
-            have not been exhausted, the connection will be restablished
+            have not been exhausted, the connection will be restablished.
+            If set to -1, no timeout will occur.
         kwargs: dict
             if a string is used for the feed, kwargs will be passed to the
             newly instantiated object
@@ -154,6 +157,9 @@ class FeedHandler:
             LOG.error("Unhandled exception", exc_info=True)
 
     async def _watch(self, feed_id, websocket):
+        if self.timeout[feed_id] == -1:
+            return
+
         while websocket.open:
             if self.last_msg[feed_id]:
                 if time() - self.last_msg[feed_id] > self.timeout[feed_id]: