+++ cryptofeed/standards.py
@@ -12,15 +12,11 @@ data channel names
 import logging
 import pandas as pd
 
-from cryptofeed.defines import (L2_BOOK, L3_BOOK, TRADES, TICKER, OPEN_INTEREST, VOLUME, FUNDING,
-                                UNSUPPORTED, BITFINEX, GEMINI, BITMAX,
-                                POLONIEX, HITBTC, BITSTAMP, COINBASE, BITMEX, KRAKEN,
-                                KRAKEN_FUTURES, BINANCE, EXX, HUOBI, HUOBI_DM, OKCOIN,
-                                OKEX, COINBENE, BYBIT, FTX, TRADES_SWAP, TICKER_SWAP, L2_BOOK_SWAP,
-                                TRADES_FUTURES, TICKER_FUTURES, L2_BOOK_FUTURES,
-                                LIMIT, MARKET, FILL_OR_KILL, IMMEDIATE_OR_CANCEL, MAKER_OR_CANCEL,
-                                DERIBIT, BITTREX, BITCOINCOM, BINANCE_US,
-                                BINANCE_JERSEY, BINANCE_FUTURES, UPBIT, BLOCKCHAIN)
+from cryptofeed.defines import (L2_BOOK, L3_BOOK, TRADES, TICKER, OPEN_INTEREST, VOLUME, FUNDING, LIQUIDATIONS, UNSUPPORTED, BITFINEX, GEMINI, BITMAX,
+                                POLONIEX, HITBTC, BITSTAMP, COINBASE, BITMEX, KRAKEN, KRAKEN_FUTURES, BINANCE, EXX, HUOBI, HUOBI_DM, OKCOIN,
+                                OKEX, COINBENE, BYBIT, FTX, TRADES_SWAP, TICKER_SWAP, L2_BOOK_SWAP, TRADES_FUTURES, TICKER_FUTURES, L2_BOOK_FUTURES,
+                                LIMIT, MARKET, FILL_OR_KILL, IMMEDIATE_OR_CANCEL, MAKER_OR_CANCEL, DERIBIT, BITTREX, BITCOINCOM, BINANCE_US,
+                                BINANCE_JERSEY, BINANCE_FUTURES, UPBIT, DSX, BLOCKCHAIN)
 from cryptofeed.pairs import gen_pairs
 from cryptofeed.exceptions import UnsupportedTradingPair, UnsupportedDataFeed, UnsupportedTradingOption
 
@@ -70,7 +66,7 @@ def pair_exchange_to_std(pair):
 
 
 def timestamp_normalize(exchange, ts):
-    if exchange in {BITMEX, COINBASE, HITBTC, OKCOIN, OKEX, BYBIT, FTX, BITCOINCOM, BLOCKCHAIN}:
+    if exchange in {BITMEX, COINBASE, HITBTC, OKCOIN, OKEX, BYBIT, FTX, BITCOINCOM, DSX, BLOCKCHAIN}:
         return pd.Timestamp(ts).timestamp()
     elif exchange in {HUOBI, HUOBI_DM, BITFINEX, COINBENE, DERIBIT, BINANCE, BINANCE_US, BINANCE_JERSEY, BINANCE_FUTURES, GEMINI, BITTREX, BITMAX, KRAKEN_FUTURES, UPBIT}:
         return ts / 1000.0
@@ -85,6 +81,7 @@ _feed_to_exchange_map = {
         POLONIEX: L2_BOOK,
         HITBTC: 'subscribeOrderbook',
         COINBASE: 'level2',
+        DSX: 'subscribeOrderbook',
         BITMEX: 'orderBookL2',
         BITSTAMP: 'diff_order_book',
         KRAKEN: 'book',
@@ -101,7 +98,7 @@ _feed_to_exchange_map = {
         OKEX: 'spot/depth_l2_tbt',
         COINBENE: L2_BOOK,
         DERIBIT: 'book',
-        BYBIT: 'order_book_25L1',
+        BYBIT: 'orderBookL2_25',
         FTX: 'orderbook',
         GEMINI: L2_BOOK,
         BITTREX: 'SubscribeToExchangeDeltas',
@@ -112,6 +109,7 @@ _feed_to_exchange_map = {
     L3_BOOK: {
         BITFINEX: 'book-R0-F0-100',
         BITSTAMP: 'detail_order_book',
+        DSX: UNSUPPORTED,
         HITBTC: UNSUPPORTED,
         COINBASE: 'full',
         BITMEX: UNSUPPORTED,
@@ -138,6 +136,7 @@ _feed_to_exchange_map = {
     TRADES: {
         POLONIEX: TRADES,
         HITBTC: 'subscribeTrades',
+        DSX: 'subscribeTrades',
         BITSTAMP: 'live_trades',
         BITFINEX: 'trades',
         COINBASE: 'matches',
@@ -167,6 +166,7 @@ _feed_to_exchange_map = {
     TICKER: {
         POLONIEX: 1002,
         HITBTC: 'subscribeTicker',
+        DSX : 'subscribeTicker',
         BITFINEX: 'ticker',
         BITSTAMP: UNSUPPORTED,
         COINBASE: 'ticker',
@@ -225,6 +225,9 @@ _feed_to_exchange_map = {
         BITMEX: 'instrument',
         KRAKEN_FUTURES: 'ticker',
         DERIBIT: 'ticker'
+    },
+    LIQUIDATIONS: {
+        BITMEX: 'liquidation'
     }
 }