+++ CHANGES.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### 1.4.1
+  * Feature: Support for disabling timeouts on feeds
+  * Bugfix: #224 Ignore newly added trading pairs in Poloniex while running
+  * Feature: New exchange, DSX
+  * Bugfix: Bybit updated their API, websocket subscription to L2 book data needed to be updated
+  * Bugfix: Deribit subscription condensed into a single message to avoid issues with rate limit
+  * Bugfix: Funding interval for bitmex not converted to integer
+
 ### 1.3.1 (2020-03-17)
   * Feature: Add missing update detection to orderbooks in Binance
   * Feature: REST support for FTX